Move AnimationCalculator to animation/
