+++ src/animation/animation_calculator.h
